+++ source/interface.py
@@ -8,6 +8,7 @@ import os
 import gi
 from source.button import ButtonHandler
 from source.displayfile import DisplayFileHandler
+from source.editor import EditorHandler
 from source.viewport import ViewportHandler
 
 gi.require_version("Gtk", "3.0")
@@ -23,31 +24,48 @@ class MainWindow(Gtk.Window):
 
     __gtype_name__ = "MainWindow"  # Nome da janela principal
 
-    # Atributos ---------------------------------------------------------------
+    # Atributos privados
     viewport_handler: ViewportHandler
+    display_file_handler: DisplayFileHandler
+    editor_handler: EditorHandler
+
+    # A ser visto
     pen_button_handler: ButtonHandler
-    line_button_handler: ButtonHandler
     width_button_handler: ButtonHandler
     color_button_handler: ButtonHandler
-    display_file_handler: DisplayFileHandler
 
-    # Obtém os widgets a serem conectados com seus handlers
+    # Os trilhões de widgets vão aqui :p
+    # Widgets
     viewport_drawing_area: Gtk.DrawingArea = Gtk.Template.Child()
-    pen_button: Gtk.Button = Gtk.Template.Child()
+    file_button: Gtk.MenuItem = Gtk.Template.Child()
+    point_button: Gtk.Button = Gtk.Template.Child()
     line_button: Gtk.Button = Gtk.Template.Child()
+    triangle_button: Gtk.Button = Gtk.Template.Child()
+    rectangle_button: Gtk.Button = Gtk.Template.Child()
+    polygon_button: Gtk.Button = Gtk.Template.Child()
+    clear_button: Gtk.Button = Gtk.Template.Child()
+    mode_label: Gtk.Label = Gtk.Template.Child()
     width_button: Gtk.SpinButton = Gtk.Template.Child()
     color_button: Gtk.ColorButton = Gtk.Template.Child()
 
-    # Construtor --------------------------------------------------------------
     def __init__(self) -> None:
 
         super().__init__()
-        self.display_file_handler = DisplayFileHandler()
-        self.viewport_handler = ViewportHandler(self.viewport_drawing_area, self.display_file_handler)
 
-        self.pen_button_handler = ButtonHandler(self.pen_button, "pen_button", self.viewport_handler)
-        self.line_button_handler = ButtonHandler(self.line_button, "line_button", self.viewport_handler)
-        self.width_button_handler = ButtonHandler(self.width_button, "width_button", self.viewport_handler)
-        self.color_button_handler = ButtonHandler(self.color_button, "color_button", self.viewport_handler)
-    
+        self.display_file_handler = DisplayFileHandler()
+        self.editor_handler = EditorHandler(self.display_file_handler,
+                                            self.file_button,
+                                            self.point_button,
+                                            self.line_button,
+                                            self.triangle_button,
+                                            self.rectangle_button,
+                                            self.polygon_button,
+                                            self.clear_button,
+                                            self.width_button,
+                                            self.color_button,
+                                            self.mode_label)
+        self.viewport_handler = ViewportHandler(self.viewport_drawing_area,
+                                                self.display_file_handler,
+                                                self.editor_handler)
+
         self.connect("destroy", Gtk.main_quit)