+++ source/transform.py
@@ -13,13 +13,11 @@ class Transform2D():
     Transformada.
     '''
 
-    # Atributos ---------------------------------------------------------------
-    
-    coord_list: list # lista de pontos (tuplas) que compõem o objeto
+    # Atributos públicos
+    coord_list: list  # lista de pontos (tuplas) que compõem o objeto
     rotation: float
     scale: list
-
-    t_matrix: np.array # translation matrix
+    t_matrix: np.array  # translation matrix
 
     # Construtor --------------------------------------------------------------
     def __init__(self, coord_list: list = [(0.0, 0.0)], rotation: float = 0.0, scale: tuple = (1.0, 1.0)) -> None:
@@ -27,23 +25,7 @@ class Transform2D():
         self.coord_list = coord_list
         self.rotation = rotation
         self.scale = list(scale)
-
-    # Getters e Setters --------------------------------------------------------
-    @property
-    def coord_list(self) -> list:
-        '''
-        Getter da posição.
-        '''
-
-        return self._coord_list
-
-    @coord_list.setter
-    def coord_list(self, value: list) -> None:
-        '''
-        Setter da posição.
-        '''
-
-        self._coord_list = value
+        self.t_matrix = None
 
     # Transformações -----------------------------------------------------------
     def translate(self, translation: tuple) -> list: